move files to src\data
