+++ src/data/nebulaInterfaces.ts
